+++ wo/cli/plugins/stack.py
@@ -85,6 +85,8 @@ class WOStackController(CementBaseController):
             (['--ngxblocker'],
                 dict(help='Install Nginx Ultimate Bad Bot Blocker',
                      action='store_true')),
+            (['--cheat'],
+                dict(help='Install cheat.sh', action='store_true')),
             (['--force'],
                 dict(help='Force install/remove/purge without prompt',
                      action='store_true')),
@@ -116,6 +118,7 @@ class WOStackController(CementBaseController):
                 (not pargs.adminer) and (not pargs.utils) and
                 (not pargs.redis) and (not pargs.proftpd) and
                 (not pargs.extplorer) and (not pargs.clamav) and
+                (not pargs.cheat) and
                 (not pargs.ufw) and (not pargs.ngxblocker) and
                 (not pargs.phpredisadmin) and (not pargs.sendmail) and
                     (not pargs.php73)):
@@ -147,6 +150,7 @@ class WOStackController(CementBaseController):
                 pargs.dashboard = True
                 pargs.phpredisadmin = True
                 pargs.extplorer = True
+                pargs.cheat = True
 
             if pargs.security:
                 pargs.fail2ban = True
@@ -426,6 +430,7 @@ class WOStackController(CementBaseController):
                     Log.debug(self, "eXtplorer is already installed")
                     Log.info(self, "eXtplorer is already installed")
 
+            # ultimate ngx_blocker
             if pargs.ngxblocker:
                 if not os.path.isdir('/etc/nginx/bots.d'):
                     Log.debug(self, "Setting packages variable for ngxblocker")
@@ -438,6 +443,21 @@ class WOStackController(CementBaseController):
                 else:
                     Log.debug(self, "ngxblocker is already installed")
                     Log.info(self, "ngxblocker is already installed")
+
+            # cheat.sh
+            if pargs.cheat:
+                if ((not os.path.exists('/usr/local/bin/cht.sh')) and
+                        (not os.path.exists('/usr/bin/cht.sh'))):
+                    Log.debug(self, 'Setting packages variable for cheat.sh')
+                    packages = packages + [[
+                        "https://raw.githubusercontent.com/chubin/cheat.sh"
+                        "/master/share/cht.sh.txt",
+                        "/usr/local/bin/cht.sh",
+                        "cheat.sh"],
+                        ["https://raw.githubusercontent.com/chubin/cheat.sh"
+                         "/master/share/bash_completion.txt",
+                         "/etc/bash_completion.d/cht.sh",
+                         "cheat.sh bash_completion"]]
 
             # UTILS
             if pargs.utils:
@@ -536,6 +556,7 @@ class WOStackController(CementBaseController):
                 (not pargs.adminer) and (not pargs.utils) and
                 (not pargs.redis) and (not pargs.proftpd) and
                 (not pargs.extplorer) and (not pargs.clamav) and
+                (not pargs.cheat) and
                 (not pargs.ufw) and (not pargs.ngxblocker) and
                 (not pargs.phpredisadmin) and (not pargs.sendmail) and
                 (not pargs.php73)):
@@ -565,6 +586,7 @@ class WOStackController(CementBaseController):
             pargs.utils = True
             pargs.netdata = True
             pargs.mysqltuner = True
+            pargs.cheat = True
 
         if pargs.security:
             pargs.fail2ban = True
@@ -673,6 +695,14 @@ class WOStackController(CementBaseController):
             if os.path.isfile('/usr/bin/mysqltuner'):
                 Log.debug(self, "Removing packages for MySQLTuner ")
                 packages = packages + ['/usr/bin/mysqltuner']
+
+        # cheat.sh
+        if pargs.cheat:
+            if os.path.isfile('/usr/local/bin/cht.sh'):
+                Log.debug(self, "Removing packages for cheat.sh ")
+                packages = packages + [
+                    '/usr/local/bin/cht.sh', '/usr/local/bin/cheat',
+                    '/etc/bash_completion.d/cht.sh']
 
         # PHPREDISADMIN
         if pargs.phpredisadmin:
@@ -794,6 +824,7 @@ class WOStackController(CementBaseController):
                 (not pargs.adminer) and (not pargs.utils) and
                 (not pargs.redis) and (not pargs.proftpd) and
                 (not pargs.extplorer) and (not pargs.clamav) and
+                (not pargs.cheat) and
                 (not pargs.ufw) and (not pargs.ngxblocker) and
                 (not pargs.phpredisadmin) and (not pargs.sendmail) and
                 (not pargs.php73)):
@@ -823,6 +854,7 @@ class WOStackController(CementBaseController):
             pargs.composer = True
             pargs.netdata = True
             pargs.mysqltuner = True
+            pargs.cheat = True
 
         if pargs.security:
             pargs.fail2ban = True
@@ -938,6 +970,14 @@ class WOStackController(CementBaseController):
             if os.path.isfile('/usr/bin/mysqltuner'):
                 Log.debug(self, "Removing packages for MySQLTuner ")
                 packages = packages + ['/usr/bin/mysqltuner']
+
+        # cheat.sh
+        if pargs.cheat:
+            if os.path.isfile('/usr/local/bin/cht.sh'):
+                Log.debug(self, "Removing packages for cheat.sh ")
+                packages = packages + [
+                    '/usr/local/bin/cht.sh', '/usr/local/bin/cheat',
+                    '/etc/bash_completion.d/cht.sh']
 
         # PHPREDISADMIN
         if pargs.phpredisadmin: